refactor(desktop): forward sort scope through the result tables
