+++ apps/desktop/src/renderer/src/components/editable-data-table.tsx
@@ -37,8 +37,9 @@ import { useSettingsStore } from '@/stores/settings-store'
 import { useMaskingStore } from '@/stores/masking-store'
 import { PaginationControls } from '@/components/pagination-controls'
 import { SmartFilterBar, chipMatchesRow, type FilterChip } from '@/components/smart-filter-bar'
-import { SmartSortBar } from '@/components/smart-sort-bar'
+import { SmartSortBar } from '@/components/sort/smart-sort-bar'
 import { applySorts, toggleColumnSort, type SortChip } from '@/lib/sort-model'
+import type { SortScope } from '@/lib/sort-scope'
 import type { ColumnInfo, ConnectionConfig, EditContext, ForeignKeyInfo } from '@data-peek/shared'
 import {
   flexRender,
@@ -112,7 +113,18 @@ interface EditableDataTableProps<TData> {
   connection?: ConnectionConfig | null
   onFiltersChange?: (filters: DataTableFilter[]) => void
   onSortingChange?: (sorting: DataTableSort[]) => void
+  /** Push the active filters into the query's WHERE clause. Filter bar only. */
   onApplyToQuery?: () => void
+  /**
+   * Whether a client-side sort covers every row, or only the loaded slice.
+   * Defaults to `complete` over `data` — correct for callers that hand the table
+   * the entire result set. Paginated callers must pass their real scope.
+   */
+  sortScope?: SortScope
+  /** True while a server-side sort re-run is debouncing or in flight. */
+  isSortingOnServer?: boolean
+  /** Rewrite ORDER BY and re-run so the sort covers rows not yet loaded. */
+  onSortWholeSet?: () => void
   onPageSizeChange?: (size: number) => void
   onForeignKeyClick?: (foreignKey: ForeignKeyInfo, value: unknown) => void
   onForeignKeyOpenTab?: (foreignKey: ForeignKeyInfo, value: unknown) => void
@@ -224,6 +236,9 @@ export function EditableDataTable<TData extends Record<string, unknown>>({
   onSortingChange,
   onPageSizeChange,
   onApplyToQuery,
+  sortScope,
+  isSortingOnServer,
+  onSortWholeSet,
   onForeignKeyClick,
   onForeignKeyOpenTab,
   onColumnStatsClick,
@@ -1172,7 +1187,9 @@ export function EditableDataTable<TData extends Record<string, unknown>>({
             columns={columnDefs}
             chips={sortChips}
             onChipsChange={setSortChips}
-            onApplyToQuery={onApplyToQuery}
+            scope={sortScope ?? { kind: 'complete', rows: data.length }}
+            isSortingOnServer={isSortingOnServer}
+            onSortWholeSet={onSortWholeSet}
           />
           {canEdit && (
             <div className="flex items-center px-2 py-1 border-b border-border/30">
